+++ src/main/java/vn/anhtuan/demoAPI/POJO/ChoicePOJO.java
@@ -3,7 +3,7 @@ package vn.anhtuan.demoAPI.POJO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ChoicePOJO {
-    private Integer id; // Thêm trường id
+    private Long id; // Đổi từ Integer sang Long
     private String content;
 
     @JsonProperty("is_correct")
@@ -12,15 +12,15 @@ public class ChoicePOJO {
     // Constructors
     public ChoicePOJO() {}
 
-    public ChoicePOJO(Integer id, String content, Boolean isCorrect) {
+    public ChoicePOJO(Long id, String content, Boolean isCorrect) {
         this.id = id;
         this.content = content;
         this.isCorrect = isCorrect;
     }
 
     // Getters & Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getContent() { return content; }
     public void setContent(String content) { this.content = content; }
