+++ src/main/java/vn/anhtuan/demoAPI/POJO/QuizSubmissionPOJO.java
@@ -5,13 +5,13 @@ import java.util.Map;
 
 public class QuizSubmissionPOJO {
     private Long userId;
-    private Map<Integer, List<Integer>> answers;
+    private Map<Long, List<Long>> answers;
     private Integer durationSeconds; // Thêm trường duration
 
     public QuizSubmissionPOJO() {
     }
 
-    public QuizSubmissionPOJO(Long userId, Map<Integer, List<Integer>> answers, Integer durationSeconds) {
+    public QuizSubmissionPOJO(Long userId, Map<Long, List<Long>> answers, Integer durationSeconds) {
         this.userId = userId;
         this.answers = answers;
         this.durationSeconds = durationSeconds;
@@ -26,11 +26,11 @@ public class QuizSubmissionPOJO {
         this.userId = userId;
     }
 
-    public Map<Integer, List<Integer>> getAnswers() {
+    public Map<Long, List<Long>> getAnswers() {
         return answers;
     }
 
-    public void setAnswers(Map<Integer, List<Integer>> answers) {
+    public void setAnswers(Map<Long, List<Long>> answers) {
         this.answers = answers;
     }
