+++ src/main/java/vn/anhtuan/demoAPI/Repository/QuestionRepository.java
@@ -9,9 +9,9 @@ import vn.anhtuan.demoAPI.Entity.Question;
 import java.util.List;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
-    List<Question> findByQuizId(Integer quizId);
+public interface QuestionRepository extends JpaRepository<Question, Long> { // Đổi Integer -> Long
+    List<Question> findByQuizId(Long quizId); // Đổi Integer -> Long
 
     @Query("SELECT q FROM Question q WHERE q.quiz.id = :quizId ORDER BY q.id")
-    List<Question> findQuestionsByQuizIdOrdered(@Param("quizId") Integer quizId);
+    List<Question> findQuestionsByQuizIdOrdered(@Param("quizId") Long quizId); // Đổi Integer -> Long
 }
